created SchduleItemNG component; fixed its usage; removed schedule download options
